Move build command to own file
